WIP: desenvolvimento de lib para utilização de kafka com spring: - adicionando fluxos de consumers - criando beans para instanciação de consumersFactorys de cada consumer - ajustando docker-compose.yaml Confluent para o mais recente - ajustando pacotes
