+++ lib--teste/src/main/java/com/lucassilvs/libteste/lib/properties/authentication/KafkaAuthProperties.java
@@ -1,4 +1,4 @@
-package com.lucassilvs.libteste.lib.authentication;
+package com.lucassilvs.libteste.lib.properties.authentication;
 
 import jakarta.annotation.Nonnull;
 import lombok.Data;
