+++ lib--teste/src/main/java/com/lucassilvs/libteste/lib/properties/schema_registry/SchemaRegistryConfiguration.java
@@ -1,4 +1,4 @@
-package com.lucassilvs.libteste.lib.schema_registry;
+package com.lucassilvs.libteste.lib.properties.schema_registry;
 
 import lombok.Data;
