+++ lib--teste/src/main/java/com/lucassilvs/libteste/service/producer/impl/MessagemServiceLibAvroImpl.java
@@ -1,9 +1,9 @@
-package com.lucassilvs.libteste.service.impl;
+package com.lucassilvs.libteste.service.producer.impl;
 
 
 import com.lucassilvs.kafkaproducerexemplo.gateways.kafka.UsuarioTesteAvro;
 import com.lucassilvs.libteste.request.MensagemRequest;
-import com.lucassilvs.libteste.service.MensagemService;
+import com.lucassilvs.libteste.service.producer.MensagemService;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
@@ -16,7 +16,7 @@ import java.util.Map;
 @Profile("avro")
 public class MessagemServiceLibAvroImpl implements MensagemService {
 
-    @Value("${kafka.producers.producer1.nomeTopico}")
+    @Value("${kafka.producers.producer2.nomeTopico}")
     private String nomeTopicoDefault;
