+++ lib--teste/src/main/java/com/lucassilvs/libteste/service/producer/impl/MessagemServiceLibStringImpl.java
@@ -1,8 +1,9 @@
-package com.lucassilvs.libteste.service.impl;
+package com.lucassilvs.libteste.service.producer.impl;
 
 
+import com.lucassilvs.libteste.configuration.KafkaClientsManager;
 import com.lucassilvs.libteste.request.MensagemRequest;
-import com.lucassilvs.libteste.service.MensagemService;
+import com.lucassilvs.libteste.service.producer.MensagemService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -15,17 +16,21 @@ public class MessagemServiceLibStringImpl implements MensagemService {
     @Value("${kafka.producers.producer1.nomeTopico}")
     private String nomeTopicoDefault;
 
-    private  KafkaTemplate<String, String> kafkaTemplateMap;
+    private  KafkaTemplate<String, String> producerTeste;
+
+    public MessagemServiceLibStringImpl() {
+        this.producerTeste = KafkaClientsManager.buscaProducer("producer1");
+    }
 
     public void postarMensagem(MensagemRequest mensagem) {
-        kafkaTemplateMap.send(nomeTopicoDefault, mensagem.getMensagem());
+        producerTeste.send(nomeTopicoDefault, mensagem.getMensagem());
         System.out.println("Mensagem: " + mensagem.getMensagem() + " Saldo: " + mensagem.getSaldo());
     }
 
     @Override
     public void postarMensagem(MensagemRequest mensagem, String nomeTopico) {
 
-        kafkaTemplateMap.send(nomeTopico, mensagem.getMensagem());
+        producerTeste.send(nomeTopico, mensagem.getMensagem());
         System.out.println("Mensagem: " + mensagem.getMensagem() + " Saldo: " + mensagem.getSaldo());
     }
 }
