+++ lib--teste/src/main/resources/application.yaml
@@ -12,7 +12,7 @@ kafka:
       requestTimeoutMs: 10000
       clientId: teste
     producer2:
-      nomeTopico: topico--teste2
+      nomeTopico: topico-com-schema
       keySerializer: org.apache.kafka.common.serialization.StringSerializer
       valueSerializer: io.confluent.kafka.serializers.KafkaAvroSerializer
       bootstrapServers: localhost:9092
@@ -26,5 +26,14 @@ kafka:
       keyDeserializer: org.apache.kafka.common.serialization.StringDeserializer
       valueDeserializer: org.apache.kafka.common.serialization.StringDeserializer
       bootstrapServers: localhost:9092
-      groupId: consumer-teste
+      groupId: consumer1-teste
       clientId: teste
+    consumer2:
+      nomeTopico: topico--com-schema
+      keyDeserializer: org.apache.kafka.common.serialization.StringDeserializer
+      valueDeserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+      bootstrapServers: localhost:9092
+      groupId: consumer2-teste
+      clientId: teste
+      schemaRegistry:
+        url: http://localhost:8081
